+++ mi/drive.py
@@ -1,5 +1,43 @@
 import json
+from typing import Any, Dict
+
 import requests
+from pydantic import BaseModel
+
+
+class Properties(BaseModel):
+    width: int
+    height: int
+    avgColor: str
+
+
+class Folder(BaseModel):
+    id: str
+    createdAt: str
+    name: str
+    foldersCount: int
+    filesCount: int
+    parentId: str
+    parent: Dict[str, Any]
+
+
+class File(BaseModel):
+    id: str
+    createdAt: str
+    name: str
+    type: str
+    md5: str
+    size: int
+    isSensitive: bool
+    blurhash: str
+    properties: Properties
+    url: str
+    thumbnailUrl: str
+    comment: str
+    folderId: str
+    folder: Folder
+    userId: str
+    user: Dict[str, Any]
 
 
 class Drive(object):