+++ mi/exception.py
@@ -1,0 +1,2 @@
+class CredentialRequired(Exception):
+    pass